+++ packages/adapters/source/direct-client.ts
@@ -9,7 +9,6 @@ import {
 import {
   DredgeRouter,
   MiddlewareRequest,
-  getDataType,
   useErrorMiddlewares,
   useSuccessMiddlewares,
   useValidate,