+++ packages/route/CHANGELOG.md
@@ -1,5 +1,11 @@
 # dredge-route
 
+## 0.3.0
+
+### Minor Changes
+
+- Fix few minor issues
+
 ## 0.2.0
 
 ### Minor Changes